+++ src/itinerary/Itinerary.tsx
@@ -10,8 +10,8 @@ import { A11yProps, pickA11yProps } from '../_utils/interfaces'
 import { Place } from '../_utils/place'
 import { Bullet, BulletTypes } from '../bullet'
 import { SpacingDivider } from '../divider/spacingDivider'
-import { SubHeader } from '../subHeader'
 import { Text, TextDisplayType, TextTagType } from '../text'
+import { TextSubHeaderStrong } from '../typography/subHeaderStrong'
 import { StyledItinerary } from './Itinerary.style'
 
 export type ItineraryProps = A11yProps &
@@ -188,7 +188,9 @@ export const Itinerary = (props: ItineraryProps) => {
     <StyledItinerary className={cc(['kirk-itinerary-root', className])} {...rootA11yProps}>
       {isNonEmptyString(headline) && (
         <Fragment>
-          <SubHeader>{headline}</SubHeader>
+          <h2 className="kirk-itinerary--headline">
+            <TextSubHeaderStrong>{headline}</TextSubHeaderStrong>
+          </h2>
           <SpacingDivider />
         </Fragment>
       )}